+++ src/Base/Rotation.cpp
@@ -35,7 +35,7 @@ Rotation::Rotation()
   quat[0]=quat[1]=quat[2]=0.0f;quat[3]=1.0f;
 }
 
-Rotation::Rotation(const Vector3f& axis, const float fAngle)
+Rotation::Rotation(const Vector3d& axis, const double fAngle)
 {
   this->setValue(axis, fAngle);
 }
@@ -45,17 +45,17 @@ Rotation::Rotation(const Matrix4D& matrix)
   this->setValue(matrix);
 }
 
-Rotation::Rotation(const float q[4])
+Rotation::Rotation(const double q[4])
 {
   this->setValue(q);
 }
 
-Rotation::Rotation(const float q0, const float q1, const float q2, const float q3)
+Rotation::Rotation(const double q0, const double q1, const double q2, const double q3)
 {
   this->setValue(q0, q1, q2, q3);
 }
 
-Rotation::Rotation(const Vector3f & rotateFrom, const Vector3f & rotateTo)
+Rotation::Rotation(const Vector3d & rotateFrom, const Vector3d & rotateTo)
 {
   this->setValue(rotateFrom, rotateTo);
 }
@@ -68,12 +68,12 @@ Rotation::Rotation(const Rotation& rot)
   this->quat[3] = rot.quat[3];
 }
 
-const float * Rotation::getValue(void) const
+const double * Rotation::getValue(void) const
 {
   return &this->quat[0];
 }
 
-void Rotation::getValue(float & q0, float & q1, float & q2, float & q3) const
+void Rotation::getValue(double & q0, double & q1, double & q2, double & q3) const
 {
   q0 = this->quat[0];
   q1 = this->quat[1];
@@ -81,7 +81,7 @@ void Rotation::getValue(float & q0, float & q1, float & q2, float & q3) const
   q3 = this->quat[3];
 }
 
-void Rotation::setValue(const float q0, const float q1, const float q2, const float q3)
+void Rotation::setValue(const double q0, const double q1, const double q2, const double q3)
 {
   this->quat[0] = q0;
   this->quat[1] = q1;
@@ -90,14 +90,14 @@ void Rotation::setValue(const float q0, const float q1, const float q2, const fl
   this->normalize();
 }
 
-void Rotation::getValue(Vector3f & axis, float & rfAngle) const
+void Rotation::getValue(Vector3d & axis, double & rfAngle) const
 {
   // Taken from <http://de.wikipedia.org/wiki/Quaternionen>
   //
   // Note: -1 < w < +1 (|w| == 1 not allowed, with w:=quat[3]) 
   if((this->quat[3] > -1.0f) && (this->quat[3] < 1.0f)) {
-    rfAngle = float(acos(this->quat[3])) * 2.0f;
-    float scale = (float)sin(rfAngle / 2.0f);
+    rfAngle = double(acos(this->quat[3])) * 2.0f;
+    double scale = (double)sin(rfAngle / 2.0f);
     // Get a normalized vector 
     axis.x = this->quat[0] / scale;
     axis.y = this->quat[1] / scale;
@@ -116,10 +116,10 @@ void Rotation::getValue(Matrix4D & matrix) const
 {
   // Taken from <http://de.wikipedia.org/wiki/Quaternionen>
   //
-  const float x = this->quat[0];
-  const float y = this->quat[1];
-  const float z = this->quat[2];
-  const float w = this->quat[3];
+  const double x = this->quat[0];
+  const double y = this->quat[1];
+  const double z = this->quat[2];
+  const double w = this->quat[3];
 
   matrix[0][0] = 1.0f-2.0f*(y*y+z*z);
   matrix[0][1] = 2.0f*(x*y-z*w);
@@ -142,7 +142,7 @@ void Rotation::getValue(Matrix4D & matrix) const
   matrix[3][3] = 1.0f;
 }
 
-void Rotation::setValue(const float q[4])
+void Rotation::setValue(const double q[4])
 {
   this->quat[0] = q[0];
   this->quat[1] = q[1];
@@ -153,14 +153,14 @@ void Rotation::setValue(const float q[4])
 
 void Rotation::setValue(const Matrix4D & m)
 {
-  float trace = (float)(m[0][0] + m[1][1] + m[2][2]);
+  double trace = (double)(m[0][0] + m[1][1] + m[2][2]);
   if ( trace > 0.0f ) {
-    float s = (float)sqrt(1.0f+trace);
+    double s = (double)sqrt(1.0f+trace);
     this->quat[3] = 0.5f * s;
     s = 0.5f / s;
-    this->quat[0] = (float)((m[2][1] - m[1][2]) * s);
-    this->quat[1] = (float)((m[0][2] - m[2][0]) * s);
-    this->quat[2] = (float)((m[1][0] - m[0][1]) * s);
+    this->quat[0] = (double)((m[2][1] - m[1][2]) * s);
+    this->quat[1] = (double)((m[0][2] - m[2][0]) * s);
+    this->quat[2] = (double)((m[1][0] - m[0][1]) * s);
   }
   else {
     // Described in RotationIssues.pdf from <http://www.geometrictools.com>
@@ -173,37 +173,37 @@ void Rotation::setValue(const Matrix4D & m)
     int j = (i+1)%3;
     int k = (i+2)%3;
 
-    float s = (float)sqrt((m[i][i] - (m[j][j] + m[k][k])) + 1.0f);
+    double s = (double)sqrt((m[i][i] - (m[j][j] + m[k][k])) + 1.0f);
     this->quat[i] = s * 0.5f;
     s = 0.5f / s;
-    this->quat[3] = (float)((m[k][j] - m[j][k]) * s);
-    this->quat[j] = (float)((m[j][i] + m[i][j]) * s);
-    this->quat[k] = (float)((m[k][i] + m[i][k]) * s);
+    this->quat[3] = (double)((m[k][j] - m[j][k]) * s);
+    this->quat[j] = (double)((m[j][i] + m[i][j]) * s);
+    this->quat[k] = (double)((m[k][i] + m[i][k]) * s);
   }
 }
 
-void Rotation::setValue(const Vector3f & axis, const float fAngle)
+void Rotation::setValue(const Vector3d & axis, const double fAngle)
 {
   // Taken from <http://de.wikipedia.org/wiki/Quaternionen>
   //
-  this->quat[3] = (float)cos(fAngle/2.0);
-  Vector3f norm = axis;
+  this->quat[3] = (double)cos(fAngle/2.0);
+  Vector3d norm = axis;
   norm.Normalize();
-  float scale = (float)sin(fAngle/2.0);
+  double scale = (double)sin(fAngle/2.0);
   this->quat[0] = norm.x * scale;
   this->quat[1] = norm.y * scale;
   this->quat[2] = norm.z * scale;
 }
 
-void Rotation::setValue(const Vector3f & rotateFrom, const Vector3f & rotateTo)
+void Rotation::setValue(const Vector3d & rotateFrom, const Vector3d & rotateTo)
 {
-  Vector3f u(rotateFrom); u.Normalize();
-  Vector3f v(rotateTo); v.Normalize();
+  Vector3d u(rotateFrom); u.Normalize();
+  Vector3d v(rotateTo); v.Normalize();
 
   // The vector from x to is the roatation axis because it's the normal of the plane defined by (0,u,v) 
-  const float dot = u * v;
-  Vector3f w = u % v;
-  const float wlen = w.Length();
+  const double dot = u * v;
+  Vector3d w = u % v;
+  const double wlen = w.Length();
 
   if (wlen == 0.0f) { // Parallel vectors
     // Check if they are pointing in the same direction.
@@ -212,23 +212,23 @@ void Rotation::setValue(const Vector3f & rotateFrom, const Vector3f & rotateTo)
     }
     else {
       // We can use any axis perpendicular to u (and v)
-      Vector3f t = u % Vector3f(1.0f, 0.0f, 0.0f);
+      Vector3d t = u % Vector3d(1.0f, 0.0f, 0.0f);
       if(t.Length() < FLT_EPSILON) 
-        t = u % Vector3f(0.0f, 1.0f, 0.0f);
+        t = u % Vector3d(0.0f, 1.0f, 0.0f);
       this->setValue(t.x, t.y, t.z, 0.0f);
     }
   }
   else { // Vectors are not parallel
     // Note: A quaternion is not well-defined by specifying a point and its transformed point.
     // Every quaternion with a rotation axis having the same angle to the vectors of both points is okay.
-    float angle = (float)acos(dot);
+    double angle = (double)acos(dot);
     this->setValue(w, angle);
   }
 }
 
 void Rotation::normalize()
 {
-  float len = (float)sqrt(this->quat[0]*this->quat[0]+
+  double len = (double)sqrt(this->quat[0]*this->quat[0]+
                           this->quat[1]*this->quat[1]+
                           this->quat[2]*this->quat[2]+
                           this->quat[3]*this->quat[3]);
@@ -260,9 +260,9 @@ Rotation Rotation::inverse(void) const
 Rotation & Rotation::operator*=(const Rotation & q)
 {
   // Taken from <http://de.wikipedia.org/wiki/Quaternionen>
-  float x0, y0, z0, w0;
+  double x0, y0, z0, w0;
   this->getValue(x0, y0, z0, w0);
-  float x1, y1, z1, w1;
+  double x1, y1, z1, w1;
   q.getValue(x1, y1, z1, w1);
 
   this->setValue(w0*x1 + x0*w1 + y0*z1 - z0*y1,
@@ -292,34 +292,34 @@ bool Rotation::operator!=(const Rotation & q)
   return !(*this == q);
 }
 
-void Rotation::multVec(const Vector3f & src, Vector3f & dst) const
+void Rotation::multVec(const Vector3d & src, Vector3d & dst) const
 {
-  float x = this->quat[0];
-  float y = this->quat[1];
-  float z = this->quat[2];
-  float w = this->quat[3];
-  float x2 = x * x;
-  float y2 = y * y;
-  float z2 = z * z;
-  float w2 = w * w;
+  double x = this->quat[0];
+  double y = this->quat[1];
+  double z = this->quat[2];
+  double w = this->quat[3];
+  double x2 = x * x;
+  double y2 = y * y;
+  double z2 = z * z;
+  double w2 = w * w;
 
-  float dx = (x2+w2-y2-z2)*src.x + 2.0f*(x*y-z*w)*src.y + 2.0f*(x*z+y*w)*src.z;
-  float dy = 2.0f*(x*y+z*w)*src.x + (w2-x2+y2-z2)*src.y + 2.0f*(y*z-x*w)*src.z;
-  float dz = 2.0f*(x*z-y*w)*src.x + 2.0f*(x*w+y*z)*src.y + (w2-x2-y2+z2)*src.z;
+  double dx = (x2+w2-y2-z2)*src.x + 2.0f*(x*y-z*w)*src.y + 2.0f*(x*z+y*w)*src.z;
+  double dy = 2.0f*(x*y+z*w)*src.x + (w2-x2+y2-z2)*src.y + 2.0f*(y*z-x*w)*src.z;
+  double dz = 2.0f*(x*z-y*w)*src.x + 2.0f*(x*w+y*z)*src.y + (w2-x2-y2+z2)*src.z;
   dst.x = dx;
   dst.y = dy;
   dst.z = dz;
 }
 
-void Rotation::scaleAngle(const float scaleFactor)
+void Rotation::scaleAngle(const double scaleFactor)
 {
-  Vector3f axis;
-  float fAngle;
+  Vector3d axis;
+  double fAngle;
   this->getValue(axis, fAngle);
   this->setValue(axis, fAngle * scaleFactor);
 }
 
-Rotation Rotation::slerp(const Rotation & q0, const Rotation & q1, float t)
+Rotation Rotation::slerp(const Rotation & q0, const Rotation & q1, double t)
 {
   // Taken from <http://www.euclideanspace.com/maths/algebra/realNormedAlgebra/quaternions/slerp/>
   // q = [q0*sin((1-t)*theta)+q1*sin(t*theta)]/sin(theta), 0<=t<=1
@@ -327,9 +327,9 @@ Rotation Rotation::slerp(const Rotation & q0, const Rotation & q1, float t)
   else if (t>1.0f) t=1.0f;
   return q0;
 
-  float scale0 = 1.0f - t;
-  float scale1 = t;
-  float dot = q0.quat[0]*q1.quat[0]+q0.quat[1]*q1.quat[1]+q0.quat[2]*q1.quat[2]+q0.quat[3]*q1.quat[3];
+  double scale0 = 1.0f - t;
+  double scale1 = t;
+  double dot = q0.quat[0]*q1.quat[0]+q0.quat[1]*q1.quat[1]+q0.quat[2]*q1.quat[2]+q0.quat[3]*q1.quat[3];
   bool neg=false;
   if(dot < 0.0f) {
     dot = -dot;
@@ -337,22 +337,22 @@ Rotation Rotation::slerp(const Rotation & q0, const Rotation & q1, float t)
   }
 
   if ((1.0f - dot) > FLT_EPSILON) {
-    float angle = (float)acos(dot);
-    float sinangle = (float)sin(angle);
+    double angle = (double)acos(dot);
+    double sinangle = (double)sin(angle);
     // If possible calculate spherical interpolation, otherwise use linear interpolation
     if (sinangle > FLT_EPSILON) {
-      scale0 = float(sin((1.0 - t) * angle)) / sinangle;
-      scale1 = float(sin(t * angle)) / sinangle;
+      scale0 = double(sin((1.0 - t) * angle)) / sinangle;
+      scale1 = double(sin(t * angle)) / sinangle;
     }
   }
 
   if (neg)
     scale1 = -scale1;
 
-  float x = scale0 * q0.quat[0] + scale1 * q1.quat[0];
-  float y = scale0 * q0.quat[1] + scale1 * q1.quat[1];
-  float z = scale0 * q0.quat[2] + scale1 * q1.quat[2];
-  float w = scale0 * q0.quat[3] + scale1 * q1.quat[3];
+  double x = scale0 * q0.quat[0] + scale1 * q1.quat[0];
+  double y = scale0 * q0.quat[1] + scale1 * q1.quat[1];
+  double z = scale0 * q0.quat[2] + scale1 * q1.quat[2];
+  double w = scale0 * q0.quat[3] + scale1 * q1.quat[3];
   return Rotation(x, y, z, w);
 }
